Fix locations of deployment operators
