+++ datarobot_provider/_experimental/example_dags/advanced_datarobot_pipeline_jdbc_dag.py
@@ -15,11 +15,11 @@ from datarobot_provider.operators.bias_and_fairness import UpdateBiasAndFairness
 from datarobot_provider.operators.connections import GetOrCreateDataStoreOperator
 from datarobot_provider.operators.credentials import GetOrCreateCredentialOperator
 from datarobot_provider.operators.datarobot import CreateProjectOperator
-from datarobot_provider.operators.datarobot import DeployRecommendedModelOperator
-from datarobot_provider.operators.datarobot import GetFeatureDriftOperator
-from datarobot_provider.operators.datarobot import GetTargetDriftOperator
-from datarobot_provider.operators.datarobot import ScorePredictionsOperator
 from datarobot_provider.operators.datarobot import TrainModelsOperator
+from datarobot_provider.operators.deployment import DeployRecommendedModelOperator
+from datarobot_provider.operators.deployment import GetFeatureDriftOperator
+from datarobot_provider.operators.deployment import GetTargetDriftOperator
+from datarobot_provider.operators.deployment import ScorePredictionsOperator
 from datarobot_provider.operators.monitoring import GetAccuracyOperator
 from datarobot_provider.operators.monitoring import GetServiceStatsOperator
 from datarobot_provider.operators.monitoring import UpdateMonitoringSettingsOperator
