+++ datarobot_provider/_experimental/example_dags/datarobot_jdbc_batch_scoring_dag.py
@@ -37,7 +37,7 @@ from airflow.decorators import dag
 
 from datarobot_provider.operators.connections import GetOrCreateDataStoreOperator
 from datarobot_provider.operators.credentials import GetOrCreateCredentialOperator
-from datarobot_provider.operators.datarobot import ScorePredictionsOperator
+from datarobot_provider.operators.deployment import ScorePredictionsOperator
 from datarobot_provider.sensors.datarobot import ScoringCompleteSensor
