+++ datarobot_provider/_experimental/example_dags/datarobot_score_dag.py
@@ -26,7 +26,7 @@ from datetime import datetime
 
 from airflow.decorators import dag
 
-from datarobot_provider.operators.datarobot import ScorePredictionsOperator
+from datarobot_provider.operators.deployment import ScorePredictionsOperator
 from datarobot_provider.sensors.datarobot import ScoringCompleteSensor
